+++ lib/widgets/palabra-guardada/palabra-guardada-card.dart
@@ -4,7 +4,6 @@ import 'package:moblie/model/main.dart';
 import 'package:moblie/model/palabra.model.dart';
 import 'package:moblie/model/palabra_guardada.model.dart';
 import 'package:moblie/pages/palabra-single-screen.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 class PalabraGuardadaCard extends StatefulWidget {
   final PalabraGuardada palabra;