+++ lib/widgets/palabras/palabras.card.dart
@@ -15,6 +15,7 @@ class PalabraCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+
     return GestureDetector(
       onTap: () => Navigator.push(
         context,